+++ R/convert.R
@@ -85,13 +85,17 @@ filterMSFileDirs <- function(files, from)
 
 listMSFiles <- function(dirs, from)
 {
-    dirs <- normalizePath(unique(dirs), mustWork = FALSE)
+    dirs <- normalizePath(unique(dirs), mustWork = FALSE, winslash = "/")
     
     allExts <- MSFileExtensions()
     allExts <- unique(unlist(allExts[from]))
 
     files <- list.files(dirs, full.names = TRUE, pattern = paste0(paste0(".+\\.", allExts, collapse = "|"), "$"),
                         ignore.case = TRUE)
+    
+    # try to get to back to the original directory order
+    ord <- match(dirname(files), dirs)
+    files <- files[order(ord)]
 
     return(filterMSFileDirs(files, from))
 }