+++ NEWS.md
@@ -5,6 +5,7 @@
 * `generateAnalysisInfo()`: try to equalize the output and input directory order
 * Added get and plot methods for total ion chromatograms (TICs) and base peak chromatograms (BPCs) for the analysisInfo `data.frame`, `features` and `featureGroups` class as: `getTICs()`, `getBPCs()`, `plotTICs()` and `plotBPCs()`.
 * Updated PubChem transformations to 0.1.9
+* Fix: reporting compounds consensus with library algorithm would fail (issue #110)
 
 
 # patRoon 2.3.1